requirements.txt
Django==1.10.4
GDAL==2.1.0
psycopg2==2.6.2
djangorestframework==3.5.3
django-angular==0.8.4
